Seperate types and constants
