+++ packages/multichain/src/adapters/caip-permission-adapter-permittedChains.ts
@@ -3,14 +3,10 @@ import type { Hex } from '@metamask/utils';
 import { KnownCaipNamespace } from '@metamask/utils';
 
 import type { Caip25CaveatValue } from '../caip25Permission';
+import { KnownNotifications, KnownRpcMethods } from '../scope/constants';
 import { getUniqueArrayItems, mergeScopes } from '../scope/transform';
 import type { ScopesObject, ScopeString } from '../scope/types';
-import {
-  KnownNotifications,
-  KnownRpcMethods,
-  KnownWalletScopeString,
-  parseScopeString,
-} from '../scope/types';
+import { KnownWalletScopeString, parseScopeString } from '../scope/types';
 
 export const getPermittedEthChainIds = (
   caip25CaveatValue: Pick<
